Corey/testing

+++ server.js
@@ -1,21 +1,32 @@
+const logger = require('./logger');
+logger.module = 'server'
+
+var config;
+var passportSetup;
+var passport;
+var mailer;
+if (process.env.TESTING !== undefined) {
+  logger.log("RUNNING IN TEST MODE")
+  config = require('./test-config.json');
+} else {
+  config = require('./config.json');
+  passportSetup = require('./passport-setup');
+  passport = require('passport');
+  mailer = require('./email');
+}
+
 const express = require('express');
 const bodyParser = require('body-parser');
 const db = require('./database');
 const utils = require('./utils');
-const mailer = require('./email');
 const app = express();
 const fs = require('fs');
 const { Parser } = require('json2csv');
 const authRoutes = require('./routes/auth-routes');
 const settingsRoutes = require('./routes/settings-routes');
-const passportSetup = require('./passport-setup');
 const cookieSession = require('cookie-session');
-const passport = require('passport');
-const config = require('./config.json');
-const logger = require('./logger');
 const rbac = require('./rbac');
 const AccessControl = require('accesscontrol');
-logger.module = 'server';
 
 app.use(express.static(__dirname + '/public'));
 app.use(express.static(__dirname + '/views'));
@@ -26,22 +37,24 @@ app.use(bodyParser.urlencoded({
 app.use(bodyParser.json());
 app.set('view engine', 'ejs');
 
-// cookies
-app.use(cookieSession({
-  maxAge: 24 * 60 * 60 * 1000,
-  keys: [config.keys.session.cookieKey]
-}));
+if (process.env.TESTING == undefined) {
+  // cookies
+  app.use(cookieSession({
+    maxAge: 24 * 60 * 60 * 1000,
+    keys: [config.keys.session.cookieKey]
+  }));
 
-// initialize passport
-app.use(passport.initialize());
-app.use(passport.session());
+  // initialize passport
+  app.use(passport.initialize());
+  app.use(passport.session());
+}
 
 // setup routes
 app.use('/auth', authRoutes);
 app.use('/settings', settingsRoutes);
 
 const authCheck = (req, res, next) => {
-  if(!req.user){
+  if(!req.user && process.env.TESTING == undefined){
       // if not logged in
       res.redirect('/auth/google');
   } else {
@@ -52,33 +65,37 @@ const authCheck = (req, res, next) => {
 
 const permCheck = function (resource, func) {
   return async function(req, res, next) {
-    const grantInfo = ((await db.queryRoleByID(req.user.roleID))[0][0]).grantInfo;
-    const role = (Object.keys(JSON.parse(grantInfo)))[0];
-    var perm = {granted : false};
-    var ac = await rbac.getRolesFromDb();
-    switch(func) {
-      case 'create':
-        perm = ac.can(role).createAny(resource);
-        break;
-      case 'read':
-        perm = ac.can(role).readAny(resource);
-        break;   
-      case 'update':
-        perm = ac.can(role).updateAny(resource);
-        break;
-      case 'delete':
-        perm = ac.can(role).deleteAny(resource);
-        break;
-      default:
-        break;
-    }
-    if(perm.granted){
-      logger.log(`${req.user.firstName} ${req.user.lastName} was granted access to ${resource} for ${func}`)
+    if (process.env.TESTING == undefined) {
+      const grantInfo = ((await db.queryRoleByID(req.user.roleID))[0][0]).grantInfo;
+      const role = (Object.keys(JSON.parse(grantInfo)))[0];
+      var perm = {granted : false};
+      var ac = await rbac.getRolesFromDb();
+      switch(func) {
+        case 'create':
+          perm = ac.can(role).createAny(resource);
+          break;
+        case 'read':
+          perm = ac.can(role).readAny(resource);
+          break;   
+        case 'update':
+          perm = ac.can(role).updateAny(resource);
+          break;
+        case 'delete':
+          perm = ac.can(role).deleteAny(resource);
+          break;
+        default:
+          break;
+      }
+      if(perm.granted){
+        logger.log(`${req.user.firstName} ${req.user.lastName} was granted access to ${resource} for ${func}`)
+        next();
+      }
+      else{
+        logger.log(`${req.user.firstName} ${req.user.lastName} was denied access to ${resource} for ${func}`)
+        res.redirect('unauthorized');
+      }
+    } else {
       next();
-    }
-    else{
-      logger.log(`${req.user.firstName} ${req.user.lastName} was denied access to ${resource} for ${func}`)
-      res.redirect('unauthorized');
     }
   }
 }
@@ -321,7 +338,7 @@ app.post('/participant/comment/:eventID/:participantID', authCheck, async (req, 
 /**
  * Renders the participant check in list for the specified event
  */
-app.get('/participants/checkin/:id', authCheck, async (req, res) => {
+app.get('/event/:id/checkin', authCheck, async (req, res) => {
   res.render('participants/checkinParticipants', {participants: await db.queryParticipantsByEventID(req.params.id), event: (await db.queryEventByID(req.params.id))[0]})
 });
 
@@ -573,6 +590,4 @@ app.get('/loginFailed', async (req, res) => {
   res.render('loginFailed', {user: null});
 });
 
-app.listen(3000, function () {
-  logger.log('Listening on port 3000!', 'info');
-});
+module.exports = app;
